+++ EssentialApp/EssentialAppTests/FeedImageDataLoaderCacheDecoratorTests.swift
@@ -7,32 +7,7 @@
 
 import XCTest
 import EssentialFeed
-
-final class FeedImageDataLoaderCacheDecorator: FeedImageDataLoader {
-    let decoratee: FeedImageDataLoader
-    let cache: FeedImageDataCache
-
-    public init(decoratee: FeedImageDataLoader, cache: FeedImageDataCache) {
-        self.decoratee = decoratee
-        self.cache = cache
-    }
-
-    func loadImageData(
-        from url: URL,
-        completion: @escaping (FeedImageDataLoader.Result) -> Void
-    ) -> any EssentialFeed.FeedImageDataLoaderTask {
-        decoratee.loadImageData(from: url) { [weak self] result in
-            completion(
-                result.map { data in
-                    self?.cache.save(data, for: url) { _ in }
-
-                    return data
-                }
-            )
-        }
-    }
-
-}
+import EssentialApp
 
 final class FeedImageDataLoaderCacheDecoratorTests: XCTestCase, FeedImageDataLoaderTestCase {
